move get_botocore_config
