+++ cartography/intel/aws/__init__.py
@@ -12,10 +12,9 @@ import neo4j
 
 from . import ec2
 from . import organizations
-from .ec2.util import get_botocore_config
 from .resources import RESOURCE_FUNCTIONS
 from cartography.config import Config
-from cartography.intel.aws.util.common import parse_and_validate_aws_requested_syncs
+from cartography.intel.aws.util.common import parse_and_validate_aws_requested_syncs, get_botocore_config
 from cartography.stats import get_stats_client
 from cartography.util import merge_module_sync_metadata
 from cartography.util import run_analysis_and_ensure_deps
