+++ cartography/intel/aws/config.py
@@ -5,7 +5,7 @@ from typing import List
 import boto3
 import neo4j
 
-from cartography.intel.aws.ec2.util import get_botocore_config
+from cartography.intel.aws import get_botocore_config
 from cartography.util import aws_handle_regions
 from cartography.util import run_cleanup_job
 from cartography.util import timeit
