+++ cartography/intel/aws/ec2/key_pairs.py
@@ -5,7 +5,7 @@ from typing import List
 import boto3
 import neo4j
 
-from .util import get_botocore_config
+from .. import get_botocore_config
 from cartography.graph.job import GraphJob
 from cartography.models.aws.ec2.keypairs import EC2KeyPairSchema
 from cartography.util import aws_handle_regions
