+++ cartography/intel/aws/ec2/launch_templates.py
@@ -5,7 +5,7 @@ import boto3
 import neo4j
 from botocore.exceptions import ClientError
 
-from .util import get_botocore_config
+from .. import get_botocore_config
 from cartography.client.core.tx import load
 from cartography.graph.job import GraphJob
 from cartography.models.aws.ec2.launch_template_versions import LaunchTemplateVersionSchema
