+++ cartography/intel/aws/ec2/load_balancer_v2s.py
@@ -6,7 +6,7 @@ import boto3
 import botocore
 import neo4j
 
-from .util import get_botocore_config
+from .. import get_botocore_config
 from cartography.util import aws_handle_regions
 from cartography.util import run_cleanup_job
 from cartography.util import timeit
