+++ cartography/intel/aws/ec2/load_balancers.py
@@ -5,7 +5,7 @@ from typing import List
 import boto3
 import neo4j
 
-from .util import get_botocore_config
+from .. import get_botocore_config
 from cartography.util import aws_handle_regions
 from cartography.util import run_cleanup_job
 from cartography.util import timeit
