+++ cartography/intel/aws/ec2/tgw.py
@@ -6,7 +6,7 @@ import boto3
 import botocore.exceptions
 import neo4j
 
-from .util import get_botocore_config
+from .. import get_botocore_config
 from cartography.util import aws_handle_regions
 from cartography.util import run_cleanup_job
 from cartography.util import timeit
