--- cartography/intel/aws/ec2/util.py
@@ -1,11 +1,0 @@
-import botocore.config
-
-
-# TODO memoize this
-def get_botocore_config() -> botocore.config.Config:
-    return botocore.config.Config(
-        read_timeout=360,
-        retries={
-            'max_attempts': 10,
-        },
-    )
