+++ cartography/intel/aws/kms.py
@@ -13,7 +13,7 @@ import neo4j
 from botocore.exceptions import ClientError
 from policyuniverse.policy import Policy
 
-from cartography.intel.aws.ec2.util import get_botocore_config
+from cartography.intel.aws import get_botocore_config
 from cartography.util import aws_handle_regions
 from cartography.util import run_cleanup_job
 from cartography.util import timeit
