+++ cartography/intel/aws/securityhub.py
@@ -6,7 +6,7 @@ import boto3
 import neo4j
 from dateutil import parser
 
-from cartography.intel.aws.ec2.util import get_botocore_config
+from cartography.intel.aws import get_botocore_config
 from cartography.util import run_cleanup_job
 from cartography.util import timeit
